requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_py-xgboost-mutex=2.0=cpu_0
ase=3.22.1=pyhd8ed1ab_1
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
boost=1.74.0=py39hefe7e4c_5
boost-cpp=1.74.0=h9f4b32c_7
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
brotlipy=0.7.0=py39hb82d6ee_1003
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
cairo=1.16.0=h15b3021_1010
certifi=2021.10.8=py39hcbf5309_1
cffi=1.15.0=py39h0878f49_0
cftime=1.6.0=py39h5d4886f_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.0.4=py39hcbf5309_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py39h21b164f_0
curl=7.82.0=h789b8ee_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.28=py39h415ef7b_0
debugpy=1.5.1=py39h415ef7b_0
decorator=5.1.1=pyhd8ed1ab_0
deepchem=2.6.1=pyhd8ed1ab_0
descriptastorus=2.2.0=py_0
double-conversion=3.2.0=h0e60522_0
eigen=3.4.0=h2d74725_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
expat=2.4.7=h39d44d4_0
flask=2.0.3=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.96=hce3cb01_2
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.31.2=py39hb82d6ee_0
freetype=2.10.4=h546665d_1
future=0.18.2=py39hcbf5309_4
gettext=0.19.8.1=ha2e2712_1008
gl2ps=1.4.2=h0597ee9_0
glew=2.1.0=h39d44d4_2
greenlet=1.1.2=py39h415ef7b_1
hdf4=4.2.15=h0e5069d_3
hdf5=1.12.1=nompi_h57737ce_104
icu=69.1=h0e60522_0
idna=3.3=pyhd8ed1ab_0
imbalanced-learn=0.9.0=pyhd8ed1ab_0
importlib-metadata=4.11.3=py39hcbf5309_0
importlib_metadata=4.11.3=hd8ed1ab_0
importlib_resources=5.6.0=pyhd8ed1ab_0
intel-openmp=2022.0.0=h57928b3_3663
ipykernel=6.10.0=py39h832f523_0
ipython=8.2.0=py39hcbf5309_0
itsdangerous=2.1.2=pyhd8ed1ab_0
jbig=2.1=h8d14728_2003
jedi=0.18.1=py39hcbf5309_0
jinja2=3.1.1=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h8ffe710_0
jsoncpp=1.9.4=h2d74725_3
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.2=py39hcbf5309_0
kiwisolver=1.4.0=py39h2e07f2f_0
krb5=1.19.3=hc8ab02b_0
latexcodec=2.0.1=pyh9f0ad1d_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=13_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libcblas=3.9.0=13_win64_mkl
libcurl=7.82.0=h789b8ee_0
libdeflate=1.10=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libglib=2.70.2=h3be07f2_4
libiconv=1.16=he774522_0
liblapack=3.9.0=13_win64_mkl
libnetcdf=4.8.1=nompi_h1cc8e9d_101
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h1d00b33_2
libsodium=1.0.18=h8d14728_1
libssh2=1.10.0=h9a1e1f7_2
libtheora=1.1.1=h8d14728_1005
libtiff=4.3.0=hc4061b1_3
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
libxcb=1.13=hcd874cb_1004
libxgboost=1.5.1=cpu_h110a20b_2
libxml2=2.9.12=hf5bbc77_1
libzip=1.8.0=h519de47_1
libzlib=1.2.11=h8ffe710_1014
lightgbm=3.3.2=py39h415ef7b_0
loguru=0.5.3=py39hcbf5309_3
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py39hb82d6ee_0
matminer=0.7.6=pyhd8ed1ab_0
matplotlib-base=3.5.1=py39h581301d_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mkl=2022.0.0=h0e2418a_796
monty=2022.3.12=pyhd8ed1ab_0
mpmath=1.2.1=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
netcdf4=1.5.8=nompi_py39hf113b1f_101
networkx=2.7.1=pyhd8ed1ab_1
numpy=1.22.3=py39h6331f09_0
openjpeg=2.4.0=hb211442_1
openssl=3.0.2=h8ffe710_1
packaging=21.3=pyhd8ed1ab_0
palettable=3.3.0=py_0
pandas=1.4.1=py39h2e25243_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.2=py39haa95532_1
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=9.0.1=py39ha53f419_2
pint=0.18=pyhd8ed1ab_0
pip=22.0.4=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
plotly=5.6.0=pyhd8ed1ab_0
proj=6.2.1=h3758d61_0
prompt-toolkit=3.0.27=pyha770c72_0
psutil=5.9.0=py39hb82d6ee_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd8ed1ab_0
py-xgboost=1.5.1=cpu_py39ha538f94_2
pyaml=21.10.1=pyhd8ed1ab_0
pybtex=0.24.0=pyhd8ed1ab_2
pycairo=1.21.0=py39h1f09dad_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pymatgen=2022.3.24=py39h2e07f2f_0
pymongo=4.0.2=py39h415ef7b_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.1=py39hb82d6ee_0
pysocks=1.7.1=py39hcbf5309_4
python=3.9.12=hcf16a7b_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2022.1=pyhd8ed1ab_0
pywin32=303=py39hb82d6ee_0
pyyaml=6.0=py39hb82d6ee_3
pyzmq=22.3.0=py39he46f08e_1
rdkit=2021.09.5=py39hfadf033_0
reportlab=3.5.68=py39h12d40c5_1
requests=2.27.1=pyhd8ed1ab_0
ruamel.yaml=0.17.21=py39hb82d6ee_0
ruamel.yaml.clib=0.2.6=py39hb82d6ee_0
scikit-learn=1.0.2=py39he931e04_0
scikit-optimize=0.9.0=pyhd8ed1ab_1
scipy=1.8.0=py39hc0c34ad_1
seaborn=0.11.2=pypi_0
setuptools=61.2.0=py39hcbf5309_0
six=1.16.0=pyh6c4a22f_0
spglib=1.16.3=py39h5d4886f_0
sqlalchemy=1.4.32=py39hb82d6ee_0
sqlite=3.38.0=h2bbff1b_0
stack_data=0.2.0=pyhd8ed1ab_0
statsmodels=0.13.2=py39h2bbff1b_0
sympy=1.10.1=py39hcbf5309_0
tabulate=0.8.9=pyhd8ed1ab_0
tbb=2021.5.0=h2d74725_0
tenacity=8.0.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
tornado=6.1=py39hb82d6ee_2
tqdm=4.63.1=pyhd8ed1ab_0
traitlets=5.1.1=pyhd8ed1ab_0
tzdata=2022a=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
uncertainties=3.1.6=pyhd8ed1ab_0
unicodedata2=14.0.0=py39hb82d6ee_0
urllib3=1.26.9=pyhd8ed1ab_0
utfcpp=3.2.1=h57928b3_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
vtk=9.0.3=py39h84a6d72_201
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd8ed1ab_0
win32_setctime=1.1.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=py39hcbf5309_3
xgboost=1.5.1=cpu_py39ha538f94_2
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h8ffe710_1014
zstd=1.5.2=h6255e5f_0
